feat(data fetching inside the onMounted hook Caveat and solution): data fetching when you have Vue-Router inside your app.
Looking at the event booking app we've create before, that it's perfectly normal practice to try and fetch the data when the component is mounted. This is how we solve that before, and we just displayed the loading state before the data for a given route is actually fetched. That's a good practice that we can stick to.

- With Vue Router there is a small caveat though. If you navigate from the same route, that you navigate to a route that is using the same component for the View, what will happen is the same component would be reused, which means that the onMounted hook won't be called for this second route, which means that if you load the data inside onMounted hook and for example you navigate to another one and suddenly the params of the route do change, but the onMounted hook won't be called again.

- Add "immediate" in watcher would not only run when the value of the id parameter changes, it would also run on the initialization of this value.

When you need to load the data for a page and you want it to do that in onMounted hook, which would ne a perfectly normal assumption. In such cases, you just need to use a watcher to make it happen.